rowenc: move low-level value encoding functions to subpackage
This commit moves the lower level utilities for encoding datum values
into values to the `rowenc/valueside` package.

Release note: None
