+++ pkg/col/coldataext/datum_vec.go
@@ -18,6 +18,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/colexecerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/memsize"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
+	"github.com/cockroachdb/cockroach/pkg/sql/rowenc/valueside"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/errors"
@@ -133,7 +134,7 @@ func (dv *datumVec) Cap() int {
 // MarshalAt implements coldata.DatumVec interface.
 func (dv *datumVec) MarshalAt(appendTo []byte, i int) ([]byte, error) {
 	dv.maybeSetDNull(i)
-	return rowenc.EncodeTableValue(
+	return valueside.Encode(
 		appendTo, descpb.NoColumnID, dv.data[i], dv.scratch,
 	)
 }
@@ -141,7 +142,7 @@ func (dv *datumVec) MarshalAt(appendTo []byte, i int) ([]byte, error) {
 // UnmarshalTo implements coldata.DatumVec interface.
 func (dv *datumVec) UnmarshalTo(i int, b []byte) error {
 	var err error
-	dv.data[i], _, err = rowenc.DecodeTableValue(&dv.da, dv.t, b)
+	dv.data[i], _, err = valueside.Decode(&dv.da, dv.t, b)
 	return err
 }
