+++ pkg/col/colserde/BUILD.bazel
@@ -48,7 +48,7 @@ go_test(
         "//pkg/sql/execinfra",
         "//pkg/sql/memsize",
         "//pkg/sql/randgen",
-        "//pkg/sql/rowenc",
+        "//pkg/sql/rowenc/valueside",
         "//pkg/sql/types",
         "//pkg/testutils",
         "//pkg/util/json",
