+++ pkg/sql/colencoding/BUILD.bazel
@@ -12,8 +12,8 @@ go_library(
         "//pkg/col/coldata",
         "//pkg/roachpb:with-mocks",
         "//pkg/sql/catalog/descpb",
-        "//pkg/sql/rowenc",
         "//pkg/sql/rowenc/keyside",
+        "//pkg/sql/rowenc/valueside",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
         "//pkg/util",
@@ -36,7 +36,7 @@ go_test(
         "//pkg/col/coldataext",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/randgen",
-        "//pkg/sql/rowenc",
+        "//pkg/sql/rowenc/valueside",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
         "//pkg/util/encoding",
