+++ pkg/sql/colexec/colexectestutils/BUILD.bazel
@@ -22,7 +22,7 @@ go_library(
         "//pkg/sql/execinfrapb",
         "//pkg/sql/parser",
         "//pkg/sql/randgen",
-        "//pkg/sql/rowenc",
+        "//pkg/sql/rowenc/valueside",
         "//pkg/sql/rowexec",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
