+++ pkg/sql/randgen/BUILD.bazel
@@ -27,6 +27,7 @@ go_library(
         "//pkg/sql/parser",
         "//pkg/sql/rowenc",
         "//pkg/sql/rowenc/keyside",
+        "//pkg/sql/rowenc/valueside",
         "//pkg/sql/sem/tree",
         "//pkg/sql/stats",
         "//pkg/sql/types",
