+++ pkg/sql/rowenc/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "rowenc",
     srcs = [
-        "column_type_encoding.go",
         "encoded_datum.go",
         "index_encoding.go",
         "partition.go",
@@ -12,7 +11,6 @@ go_library(
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/rowenc",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/geo",
         "//pkg/geo/geoindex",
         "//pkg/geo/geopb",
         "//pkg/keys",
@@ -21,25 +19,21 @@ go_library(
         "//pkg/sql/catalog/colinfo",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/inverted",
-        "//pkg/sql/lex",
         "//pkg/sql/parser",
         "//pkg/sql/rowenc/keyside",
         "//pkg/sql/rowenc/rowencpb",
+        "//pkg/sql/rowenc/valueside",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sqlerrors",
         "//pkg/sql/types",
         "//pkg/util",
         "//pkg/util/encoding",
-        "//pkg/util/ipaddr",
         "//pkg/util/json",
         "//pkg/util/log",
         "//pkg/util/mon",
         "//pkg/util/protoutil",
-        "//pkg/util/timeutil/pgdate",
         "//pkg/util/unique",
-        "//pkg/util/uuid",
         "@com_github_cockroachdb_errors//:errors",
-        "@com_github_lib_pq//oid",
     ],
 )
 
@@ -47,15 +41,13 @@ go_test(
     name = "rowenc_test",
     size = "medium",
     srcs = [
-        "column_type_encoding_test.go",
         "encoded_datum_test.go",
-        "helpers_test.go",
         "index_encoding_test.go",
         "main_test.go",
         "roundtrip_format_test.go",
     ],
-    embed = [":rowenc"],
     deps = [
+        ":rowenc",
         "//pkg/keys",
         "//pkg/kv",
         "//pkg/roachpb:with-mocks",
@@ -69,22 +61,17 @@ go_test(
         "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/inverted",
         "//pkg/sql/randgen",
+        "//pkg/sql/rowenc/valueside",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
-        "//pkg/testutils",
         "//pkg/testutils/serverutils",
         "//pkg/util/encoding",
         "//pkg/util/json",
         "//pkg/util/leaktest",
         "//pkg/util/randutil",
-        "//pkg/util/timeofday",
-        "//pkg/util/timeutil",
-        "//pkg/util/timeutil/pgdate",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_apd_v2//:apd",
         "@com_github_cockroachdb_errors//:errors",
-        "@com_github_leanovate_gopter//:gopter",
-        "@com_github_leanovate_gopter//prop",
         "@com_github_stretchr_testify//require",
     ],
 )
