+++ pkg/sql/rowenc/encoded_datum.go
@@ -19,6 +19,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/colinfo"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc/keyside"
+	"github.com/cockroachdb/cockroach/pkg/sql/rowenc/valueside"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/encoding"
@@ -251,7 +252,7 @@ func (ed *EncDatum) EnsureDecoded(typ *types.T, a *tree.DatumAlloc) error {
 	case descpb.DatumEncoding_DESCENDING_KEY:
 		ed.Datum, rem, err = keyside.Decode(a, typ, ed.encoded, encoding.Descending)
 	case descpb.DatumEncoding_VALUE:
-		ed.Datum, rem, err = DecodeTableValue(a, typ, ed.encoded)
+		ed.Datum, rem, err = valueside.Decode(a, typ, ed.encoded)
 	default:
 		return errors.AssertionFailedf("unknown encoding %d", log.Safe(ed.encoding))
 	}
@@ -295,7 +296,7 @@ func (ed *EncDatum) Encode(
 	case descpb.DatumEncoding_DESCENDING_KEY:
 		return keyside.Encode(appendTo, ed.Datum, encoding.Descending)
 	case descpb.DatumEncoding_VALUE:
-		return EncodeTableValue(appendTo, descpb.NoColumnID, ed.Datum, nil /* scratch */)
+		return valueside.Encode(appendTo, descpb.NoColumnID, ed.Datum, nil /* scratch */)
 	default:
 		panic(errors.AssertionFailedf("unknown encoding requested %s", enc))
 	}
@@ -328,7 +329,7 @@ func (ed *EncDatum) Fingerprint(
 		}
 		// We must use value encodings without a column ID even if the EncDatum already
 		// is encoded with the value encoding so that the hashes are indeed unique.
-		fingerprint, err = EncodeTableValue(appendTo, descpb.NoColumnID, ed.Datum, nil /* scratch */)
+		fingerprint, err = valueside.Encode(appendTo, descpb.NoColumnID, ed.Datum, nil /* scratch */)
 	default:
 		// For values that are key encodable, using the ascending key.
 		// Note that using a value encoding will not easily work in case when
