--- pkg/sql/rowenc/helpers_test.go
@@ -1,17 +1,0 @@
-// Copyright 2020 The Cockroach Authors.
-//
-// Use of this software is governed by the Business Source License
-// included in the file licenses/BSL.txt.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0, included in the file
-// licenses/APL.txt.
-
-package rowenc
-
-// EncodeArray forwards the definition of encodeArray for testing.
-var EncodeArray = encodeArray
-
-// DecodeArray forwards the definition of decodeArray for testing.
-var DecodeArray = decodeArray
