+++ pkg/sql/scrub_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catalogkv"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
+	"github.com/cockroachdb/cockroach/pkg/sql/rowenc/valueside"
 	"github.com/cockroachdb/cockroach/pkg/sql/scrub"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
@@ -366,7 +367,7 @@ INSERT INTO t.test VALUES (10, 2);
 	// constraint.
 	values = []tree.Datum{tree.NewDInt(10), tree.NewDInt(0)}
 	// Encode the column value.
-	valueBuf, err := rowenc.EncodeTableValue(
+	valueBuf, err := valueside.Encode(
 		[]byte(nil), tableDesc.PublicColumns()[1].GetID(), values[1], []byte(nil))
 	if err != nil {
 		t.Fatalf("unexpected error: %s", err)
@@ -687,7 +688,7 @@ INSERT INTO t.test VALUES (217, 314, 1337);
 	primaryIndexKey = keys.MakeFamilyKey(primaryIndexKey, uint32(family.ID))
 
 	// Encode the second column value.
-	valueBuf, err := rowenc.EncodeTableValue(
+	valueBuf, err := valueside.Encode(
 		[]byte(nil), tableDesc.PublicColumns()[1].GetID(), values[1], []byte(nil))
 	if err != nil {
 		t.Fatalf("unexpected error: %s", err)
@@ -790,7 +791,7 @@ CREATE TABLE t.test (
 	primaryIndexKeyWithFamily := keys.MakeFamilyKey(primaryIndexKey, uint32(tableDesc.GetFamilies()[1].ID))
 
 	// Encode the second column value.
-	valueBuf, err := rowenc.EncodeTableValue(
+	valueBuf, err := valueside.Encode(
 		[]byte(nil), tableDesc.PublicColumns()[1].GetID(), values[1], []byte(nil))
 	if err != nil {
 		t.Fatalf("unexpected error: %s", err)
@@ -808,7 +809,7 @@ CREATE TABLE t.test (
 		uint32(oldTableDesc.GetFamilies()[1].ID))
 
 	// Encode the second column value.
-	valueBuf, err = rowenc.EncodeTableValue(
+	valueBuf, err = valueside.Encode(
 		[]byte(nil), tableDesc.PublicColumns()[1].GetID(), values[1], []byte(nil))
 	if err != nil {
 		t.Fatalf("unexpected error: %s", err)
@@ -895,13 +896,13 @@ CREATE TABLE t.test (k INT PRIMARY KEY, v1 INT, v2 INT);
 
 	// Encode the second column values. The second column is encoded with
 	// a garbage colIDDiff.
-	valueBuf, err := rowenc.EncodeTableValue(
+	valueBuf, err := valueside.Encode(
 		[]byte(nil), tableDesc.PublicColumns()[1].GetID(), values[1], []byte(nil))
 	if err != nil {
 		t.Fatalf("unexpected error: %s", err)
 	}
 
-	valueBuf, err = rowenc.EncodeTableValue(valueBuf, 1000, values[2], []byte(nil))
+	valueBuf, err = valueside.Encode(valueBuf, 1000, values[2], []byte(nil))
 	if err != nil {
 		t.Fatalf("unexpected error: %s", err)
 	}
