+++ pkg/sql/sqlinstance/instancestorage/BUILD.bazel
@@ -22,6 +22,7 @@ go_library(
         "//pkg/sql/catalog/systemschema",
         "//pkg/sql/row",
         "//pkg/sql/rowenc",
+        "//pkg/sql/rowenc/valueside",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sqlinstance",
         "//pkg/sql/sqlliveness",
